[UF-355] Upgrading project archetype for Errai UI and also removing old (unmaintained) component archetype
