--- uberfire-archetype/uberfire-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-api/.gitignore
@@ -1,25 +1,0 @@
-*.class
-
-# Package Files #
-*.jar
-*.war
-*.ear
-/.metadata
-/target
-/local
-/bin
-
-# Eclipse, Netbeans and IntelliJ files
-/.*
-/**/.*
-!.gitignore
-/nbproject
-*.ipr
-*.iws
-*.iml
-
-# Repository wide ignore mac DS_Store files
-.DS_Store
-
-# Bitronix transactin logs
-*.tlog
