--- uberfire-archetype/uberfire-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-api/src/main/resources/ErraiApp.properties
@@ -1,20 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-${symbol_pound}
-${symbol_pound} ErraiApp.properties
-${symbol_pound}
-${symbol_pound} Do not remove, even if empty!
-${symbol_pound}
-
-${symbol_pound} This is a marker file. When it is detected inside a JAR or at the
-${symbol_pound} top of any classpath, the subdirectories are scanned for deployable
-${symbol_pound} components. As such, all Errai application modules in a project
-${symbol_pound} should contain an ErraiApp.properties at the root of all classpaths
-${symbol_pound} that you wish to be scanned.
-${symbol_pound}
-${symbol_pound} There are also some configuration options that can be set in this
-${symbol_pound} file, although it is rarely necessary. See the documentation at
-${symbol_pound} https://docs.jboss.org/author/display/ERRAI/ErraiApp.properties
-${symbol_pound} for details.
-
