--- uberfire-archetype/uberfire-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-api/src/main/resources/__capitalizedComponentId__API.gwt.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.5.1//EN" "http://google-web-toolkit.googlecode.com/svn/tags/2.5.1/distro-source/core/src/gwt-module.dtd">
-
-<module>
-  <inherits name="org.jboss.errai.bus.ErraiBus"/>
-
-  <source path='model'/>
-  <source path='service'/>
-
-</module>
