--- uberfire-archetype/uberfire-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-client/src/main/resources/__capitalizedComponentId__Client.gwt.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.5.1//EN" "http://google-web-toolkit.googlecode.com/svn/tags/2.5.1/distro-source/core/src/gwt-module.dtd">
-
-<module>
-  <inherits name="org.uberfire.UberfireAPI"/>
-  <inherits name="org.uberfire.UberfireClientAPI"/>
-
-  <inherits name="org.gwtbootstrap3.GwtBootstrap3NoTheme"/>
-
-  <inherits name="${package}.${capitalizedComponentId}API"/>
-
-  <source path='client'/>
-
-</module>
