--- uberfire-archetype/uberfire-component-archetype/src/main/resources/archetype-resources/pom.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>${artifactId}</artifactId>
-  <packaging>pom</packaging>
-
-  <name>${capitalizedComponentId} Component</name>
-</project>
