--- uberfire-archetype/uberfire-component-archetype/src/test/resources/projects/basic/archetype.properties
@@ -1,23 +1,0 @@
-#
-# Copyright 2015 Red Hat, Inc. and/or its affiliates.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-#Thu Mar 20 20:55:34 BRT 2014
-package=com.example.template
-version=0.1-SNAPSHOT
-groupId=com.example
-artifactId=project-basic
-componentId=basic
-capitalizedComponentId=Basic
