+++ uberfire-archetype/uberfire-project-archetype/src/main/resources/META-INF/maven/archetype-metadata.xml
@@ -150,6 +150,7 @@
               <directory>src/main/java</directory>
               <includes>
                 <include>**/*.xml</include>
+                <include>**/*.html</include>
                 <include>**/*.java</include>
               </includes>
             </fileSet>
