--- uberfire-archetype/uberfire-project-archetype/src/main/resources/archetype-resources/__rootArtifactId__-showcase/__rootArtifactId__-webapp/src/main/java/client/resources/css/welcome.css
@@ -1,110 +1,0 @@
-.welcome-box {
-    background-color: #FFF;
-    border: 1px solid #CDCBCB;
-    margin-top: 50px;
-}
-
-.welcome-box h3 {
-    font-size: 34px;
-    letter-spacing: 0.5px;
-    font-weight: 100;
-    text-shadow: 0 1px 0 rgba(255, 255, 255, .7);
-}
-
-.welcome-box .welcome-top {
-    height: 54px;
-    background-color: #464141;
-    padding: 0 20px;
-    color: #fff;
-    line-height: 54px;
-}
-
-.welcome-box .welcome-body {
-    height: 425px;
-}
-
-@url getStartedBgr getStartedBgr;
-
-.welcome-box .welcome-body .get-started {
-    position: relative;
-    background: getStartedBgr no-repeat top left;
-    height: 425px;
-    text-align: center;
-    -webkit-box-sizing: border-box;
-    box-sizing: border-box;
-}
-
-.welcome-box .welcome-body .get-started h3 {
-    margin-bottom: 20px;
-}
-
-.welcome-box .welcome-body .get-started::before {
-    position: absolute;
-    height: 423px;
-    border-right: 1px solid #CDCBCB;
-    right: -15px;
-    content: '';
-    top: 1px;
-}
-
-@url gitBgr gitBgr;
-
-.welcome-box .welcome-body .git {
-    height: 212px;
-    border-bottom: 1px solid #CDCBCB;
-    margin-left: -15px;
-    -webkit-box-sizing: border-box;
-    -moz-box-sizing: border-box;
-    box-sizing: border-box;
-    background: gitBgr no-repeat top right;
-}
-
-@url downloadsBgr downloadsBgr;
-
-.welcome-box .welcome-body .downloads {
-    height: 212px;
-
-    -webkit-box-sizing: border-box;
-    -moz-box-sizing: border-box;
-    box-sizing: border-box;
-    background: downloadsBgr no-repeat top right;
-}
-
-.welcome-box .link-list a {
-    display: block;
-    padding: 0 20px;
-    background-color: #464141;
-    color: #fff;
-    text-decoration: none;
-    margin-bottom: 1px;
-    height: 34px;
-    line-height: 34px;
-    font-size: 13px;
-
-}
-
-.welcome-box .link-list a:hover {
-    background-color: #E95A35;
-}
-
-.welcome-box .git .link-list a {
-    display: block;
-    width: 170px;
-    text-align: center;
-    background-color: #6BACC6;
-}
-
-.welcome-box .git .link-list a:hover {
-    background-color: #4886B9;
-}
-
-.welcome-box .downloads .link-list a {
-    display: block;
-    width: 170px;
-    text-align: center;
-    background-color: #84C1A7;
-}
-
-.welcome-box .downloads .link-list a:hover {
-    background-color: #63A38F;
-}
