--- uberfire-archetype/uberfire-project-archetype/src/main/resources/archetype-resources/__rootArtifactId__-showcase/__rootArtifactId__-webapp/src/main/java/client/screens/HelloWorldScreen.java
@@ -1,37 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-package ${package}.client.screens;
-
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.Dependent;
-
-import com.google.gwt.core.client.GWT;
-import com.google.gwt.uibinder.client.UiBinder;
-import com.google.gwt.user.client.ui.Composite;
-import com.google.gwt.user.client.ui.Widget;
-import org.uberfire.client.annotations.WorkbenchPartTitle;
-import org.uberfire.client.annotations.WorkbenchScreen;
-
-@Dependent
-@WorkbenchScreen(identifier = "HelloWorldScreen")
-public class HelloWorldScreen extends Composite {
-
-    interface Binder
-            extends
-            UiBinder<Widget, HelloWorldScreen> {
-
-    }
-
-    private static Binder uiBinder = GWT.create( Binder.class );
-
-    @PostConstruct
-    public void init() {
-        initWidget( uiBinder.createAndBindUi( this ) );
-    }
-
-    @WorkbenchPartTitle
-    public String getTitle() {
-        return "Greetings";
-    }
-}
