--- uberfire-archetype/uberfire-project-archetype/src/main/resources/archetype-resources/__rootArtifactId__-showcase/__rootArtifactId__-webapp/src/main/java/client/screens/HelloWorldScreen.ui.xml
@@ -1,51 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-<?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
-<ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
-             xmlns:gwt="urn:import:com.google.gwt.user.client.ui">
-
-  <ui:with field='res' type='${package}.client.resources.AppResource'/>
-
-  <gwt:HTMLPanel>
-    <div class="{res.CSS.welcomeBox} container">
-      <div class="{res.CSS.welcomeTop}">
-        Learn and Discover
-        <strong>uberfire</strong>
-      </div>
-      <div class="row {res.CSS.welcomeBody}">
-        <div class="col-md-4 {res.CSS.getStarted}">
-          <div style="padding: 90px 60px 0 ;">
-            <h3>Get Started</h3>
-            <div class="{res.CSS.linkList}">
-              <a href="${symbol_pound}">Documentation</a>
-              <a href="${symbol_pound}">Get Started with tutorials</a>
-              <a href="${symbol_pound}">Help us to write documentation</a>
-            </div>
-          </div>
-        </div>
-        <div class="col-md-8">
-          <div class="{res.CSS.git}">
-            <div style="padding: 66px 60px;">
-              <div class="{res.CSS.linkList}">
-                <a href="${symbol_pound}">Fork from github</a>
-                <a href="${symbol_pound}">Send Pull Requests</a>
-              </div>
-            </div>
-          </div>
-          <div class="{res.CSS.downloads}">
-            <div style="padding: 45px 45px;">
-              <h3>Download uberfire</h3>
-              <div class="{res.CSS.linkList}">
-                <a href="${symbol_pound}">download uberfire</a>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-  </gwt:HTMLPanel>
-
-</ui:UiBinder>
