+++ uberfire-archetype/uberfire-project-archetype/src/main/resources/archetype-resources/__rootArtifactId__-showcase/__rootArtifactId__-webapp/src/main/java/client/screens/HelloWorldUberfirePresenter.java
@@ -1,0 +1,33 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+package ${package}.client.screens;
+
+import org.uberfire.client.annotations.WorkbenchPartTitle;
+import org.uberfire.client.annotations.WorkbenchPartView;
+import org.uberfire.client.annotations.WorkbenchScreen;
+import org.uberfire.client.mvp.UberElement;
+import javax.enterprise.context.Dependent;
+import javax.inject.Inject;
+
+@Dependent
+@WorkbenchScreen( identifier = "HelloWorldScreen" )
+public class HelloWorldUberfirePresenter {
+
+    public interface View extends UberElement<HelloWorldUberfirePresenter> {
+    }
+
+    @Inject
+    private View view;
+
+    @WorkbenchPartTitle
+    public String getTitle() {
+        return "HelloWorldScreen";
+    }
+
+    @WorkbenchPartView
+    public UberElement<HelloWorldUberfirePresenter> getView() {
+        return view;
+    }
+
+}
