+++ uberfire-archetype/uberfire-project-archetype/src/main/resources/archetype-resources/__rootArtifactId__-showcase/__rootArtifactId__-webapp/src/main/java/client/screens/HelloWorldUberfireView.html
@@ -1,0 +1,11 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+<div class="container" data-field="container">
+  <div class="jumbotron">
+    <h1>Welcome to Uberfire</h1>
+    <p class="lead">How about take a look on our cool tutorials and docs?</p>
+    <p><a class="btn btn-lg btn-success"
+          href="http://www.uberfireframework.org/docs/" role="button">Go for it!!!</a></p>
+  </div>
+</div>
